Began refactoring agent classes to simplify the running of the simulation. sensors refactored to prevent duplicated code
