+++ .gitignore
@@ -117,3 +117,4 @@ Visualisations/
 
 reinfLearning/
 
+OccupancyGrid/Archive/
