+++ OccupancyGrid/Agents/BaseOccupancyGridAgentconfig.ini
@@ -1,0 +1,7 @@
+
+[DIRECTORIES]
+ImageDir = D:/ReinforcementLearning/DetectSourceAgent/Data/SensorData
+AgentStateDir = D:/ReinforcementLearning/DetectSourceAgent/Analysis
+ObservationDir = D:/ReinforcementLearning/DetectSourceAgent/Observations
+MockedImageDir = D:/ReinforcementLearning/DetectSource/Data/MockData
+
